move weather tests into plugins
